Do not show items from far in the future on the dashboard
Explanation:
The logfiles do not store the year in each line, so when extracting the data,
the year is set to the current year. This is problematic when the year changes
(eg. from 31.12.2017 23:59:10 to 01.01.2018 00:00:10). If tamandua starts parsing
the logfile on 01.01.2018 00:10:00 then it will also parse data from the last year
(2017) but will set the year to the current (2018) and thus generating dates like
31.12.2018. (eg. this was written to the logfile at 31.12.2017 23:50:00, but because
the year is not stored in the logfile tamandua will just assume the current year.

This problem is solved by just limiting the end date of the request.